% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_map_data_merc.R
\docType{data}
\name{county_map_data_merc}
\alias{county_map_data_merc}
\title{Dataset for mapping U.S. counties}
\format{A data frame with 206,500 rows and 7 variables}
\usage{
data(county_map_data_merc)
}
\description{
A fortified data set that includes U.S. counties and is suitable for making maps with \code{ggplot2}.
The county FIPS codes and boundary lines from the most recent TIGER release from the U.S. Census Bureau.

\itemize{
  \item \code{long}: County longitude
  \item \code{lat}: County latitude
  \item \code{order}: Polygon order
  \item \code{hole}: hole
  \item \code{piece}: Polygon piece
  \item \code{id}: FIPS Code
  \item \code{group}: group
}
}
\details{
Dataset with the lat. / long. of county FIPS codes used for mapping county level data to a Mercator projection of the US or a single state.
To access the data directly, issue the command \code{county_map_data_merc)}.
}
\note{
Last updated 2016-05-26
}
\keyword{datasets}

